Moved and fixed integration tests
