+++ packages/jss-observable/.size-snapshot.json
@@ -1,11 +1,11 @@
 {
   "dist/jss-observable.js": {
-    "bundled": 17825,
+    "bundled": 18042,
     "minified": 5671,
     "gzipped": 2070
   },
   "dist/jss-observable.min.js": {
-    "bundled": 17825,
+    "bundled": 18042,
     "minified": 5671,
     "gzipped": 2070
   }
