+++ packages/jss-plugin-syntax-rule-value-function/.size-snapshot.json
@@ -18,5 +18,15 @@
     "bundled": 48426,
     "minified": 19395,
     "gzipped": 5022
+  },
+  "dist/jss-plugin-syntax-rule-value-function.js": {
+    "bundled": 48496,
+    "minified": 19349,
+    "gzipped": 4997
+  },
+  "dist/jss-plugin-syntax-rule-value-function.min.js": {
+    "bundled": 48496,
+    "minified": 19349,
+    "gzipped": 4997
   }
 }
