+++ packages/jss-plugin-syntax-rule-value-function/src/function-rules.test.js
@@ -133,4 +133,3 @@ describe('jss-plugin-syntax-rule-value-function: Function rules', () => {
     })
   })
 })
-
